requirements.txt
# API Contract Validator (editable install for live updates)
-e /Users/I764709/api-contract-validator

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0

# Caching
redis>=5.0.0
hiredis>=2.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0

# HTTP client for inter-service communication
httpx>=0.27.0

# Monitoring
prometheus-client>=0.20.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
email-validator>=2.1.0
